requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
idna==3.10
instaloader==4.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
requests==2.32.3
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
gunicorn
python-dotenv
